R5-G1-3 (clang-tidy narrowing tier, part 4): last qint64->double, tier now 0
One skipped display cast had drifted onto an unrelated line during the earlier
edits and was left for a manual pass; it is user_profile_backup_wizard_pages.cpp
line 1896, source->size_bytes / kBytesPerMBf, the same qint64 -> double MB
display as the rest. Made explicit with static_cast<double>. Re-running the two
narrowing checks over all 301 first-party TUs now reports 0 findings, down from
the 400 measured at the start of the tier.

Also records the narrowing-tier outcome and the two measurement gotchas
(clang-diagnostic-as-error aborting TUs; CRLF workflow scripts rejected) in
docs/CODEX_REVIEW_5_REMEDIATION.md.

clang-format + lizard clean, full Release build + ctest 225/225.

+++ src/gui/user_profile_backup_wizard_pages.cpp
@@ -1893,7 +1893,7 @@ void UserProfileBackupAppDataPage::populateTree(const QVector<AppDataSourceInfo>
             auto* item = new QTreeWidgetItem(categoryItem);
             item->setText(kAppDataColumnName, source->name);
             item->setText(kAppDataColumnPath, source->relative_path);
-            const double sizeMB = source->size_bytes / sak::kBytesPerMBf;
+            const double sizeMB = static_cast<double>(source->size_bytes) / sak::kBytesPerMBf;
             item->setText(kAppDataColumnSize,
                           QString("%1 MB").arg(sizeMB, 0, 'f', kAppDataSizeDisplayPrecision));
             item->setFlags(item->flags() | Qt::ItemIsUserCheckable);
